requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.4
mysqlclient==2.2.0
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psutil==5.9.6
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.6
